break up PolicyEngine.test.js
